# reduce amount allocation for single number decode