… README.md
… src/Hashids.net/MultipleResultsException.cs
… test/Hashids.net.test/GeneralTests.cs