Reason for change:
Almost 5x saving for single decode
After further changes by adding a optimized splitter using `Span<char>` it is only a 2x saving.
By adding a specific `GetNumberFrom` that handles a single number we avoid creating unnecessary array for `long[]` and can simply pass back `long`

I think it would greatly benefit to also actually have a `GenerateHashFrom` that takes single int/long that would mean single encode decode could be 100% allocation free.

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1620 (21H2)
AMD Ryzen 9 3950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.103
  Job-RCMHEX : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT

|               Method |        Job |            Runtime |       Mean |     Error |    StdDev | Code Size |  Gen 0 | Allocated |
|--------------------- |----------- |------------------- |-----------:|----------:|----------:|----------:|-------:|----------:|
|        DecodeNumbers | Job-RCMHEX |           .NET 6.0 |   864.4 ns |   3.99 ns |   3.74 ns |   1,359 B | 0.0181 |     152 B |
|   DecodeSingleNumber | Job-RCMHEX |           .NET 6.0 |   799.2 ns |   5.65 ns |   5.28 ns |     190 B | 0.0038 |      32 B |
```