@jetersen 2022-04-03
I may have missed the purpose of the `_guards` split in the GetNumbersFrom and whether it applies to single.