@KeterSCP 2022-04-04
The following code throws `NoResultException` in your PR if no _guards were used:

```c#
var hashids = new HashidsNet.Hashids("salt", minHashLength: 4);

var hash = hashids.Encode(123);
Console.WriteLine(hash);

var decoded = hashids.DecodeSingle(hash); // <-- Exception here
Console.WriteLine(decoded);
```