@jetersen 2022-04-04
I think I managed to find a way to actually improve Decode for multiple numbers without allocating string arrays for guards 😄