@jetersen 2022-04-04
@KeterSCP [bc5f785](https://github.com/ullmark/hashids.net/pull/65/commits/bc5f7857541791060d67476b6f8f3fc4ff1fc70b) could also apply to multiple decoding.