@jetersen 2022-04-04
```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1620 (21H2)
AMD Ryzen 9 3950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.103
  [Host]     : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  Job-AEAOHU : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT

OutlierMode=DontRemove  Runtime=.NET 6.0  MemoryRandomization=True  
```

Before: [a7478cd](https://github.com/ullmark/hashids.net/pull/65/commits/a7478cd1b8e8dbe02dde8c7f609b5d3d7388cd97)
```
|               Method |       Mean |    Error |   StdDev | Code Size |  Gen 0 | Allocated |
|--------------------- |-----------:|---------:|---------:|----------:|-------:|----------:|
|        RoundtripInts | 3,785.7 ns | 18.53 ns | 17.33 ns |      50 B | 0.0610 |     512 B |
|       RoundtripLongs | 4,287.0 ns | 18.31 ns | 17.13 ns |   3,330 B | 0.0610 |     512 B |
|         RoundtripHex | 3,725.4 ns | 22.88 ns | 21.41 ns |      49 B | 0.1602 |   1,344 B |
|         SingleNumber |   844.0 ns |  3.69 ns |  3.45 ns |     145 B | 0.0076 |      64 B |
|        DecodeNumbers |   862.9 ns |  4.09 ns |  3.82 ns |   1,359 B | 0.0181 |     152 B |
|      RoundtripSingle | 1,255.6 ns |  8.97 ns |  8.39 ns |     304 B | 0.0076 |      64 B |
| SingleNumberAsParams |   862.7 ns |  4.93 ns |  4.61 ns |   2,040 B | 0.0191 |     160 B |
```

After: [a7478cd](https://github.com/ullmark/hashids.net/pull/65/commits/a7478cd1b8e8dbe02dde8c7f609b5d3d7388cd97)
```
|               Method |       Mean |    Error |   StdDev | Code Size |  Gen 0 | Allocated |
|--------------------- |-----------:|---------:|---------:|----------:|-------:|----------:|
|        RoundtripInts | 3,827.1 ns | 25.90 ns | 24.22 ns |      50 B | 0.0305 |     264 B |
|       RoundtripLongs | 4,356.8 ns | 34.19 ns | 31.98 ns |   3,612 B | 0.0229 |     224 B |
|         RoundtripHex | 3,817.9 ns | 13.13 ns | 12.28 ns |      49 B | 0.1335 |   1,128 B |
|         SingleNumber |   846.4 ns |  5.44 ns |  5.09 ns |     145 B | 0.0076 |      64 B |
|        DecodeNumbers |   864.6 ns |  3.67 ns |  3.43 ns |   1,641 B | 0.0076 |      64 B |
|      RoundtripSingle | 1,248.2 ns |  7.87 ns |  7.36 ns |     304 B | 0.0076 |      64 B |
| SingleNumberAsParams |   855.6 ns |  3.93 ns |  3.67 ns |   2,040 B | 0.0191 |     160 B |
```