@KeterSCP 2022-04-04
New test for `long.MaxValue` fails in this PR

![image](https://user-images.githubusercontent.com/38229504/161634231-0d53b741-b2e6-4ae4-b2e5-b321f670c961.png)