@KeterSCP 2022-04-04
Hmm, the following code produces different exception for `main` and PR branches:

```c#
var hashids = new HashidsNet.Hashids("salt", minHashLength: 0, alphabet: "0123456789ABCDEF");

var hash = hashids.EncodeLong(long.MaxValue);
Console.WriteLine(hash);

var decoded = hashids.DecodeSingleLong(hash + "1"); // HashidsNet.NoResultException on main branch and HashidsNet.MultipleResultsException on PR branch
Console.WriteLine(decoded);
```