@jetersen 2022-04-05
@KeterSCP That's true because I decided to check for separators ahead of time. Previous code was doing after the decode had returned multiple numbers.

So that's a toss up and this behavior was only added in 1.5.0 so which behavior is more correct?