@KeterSCP 2022-04-05
> @KeterSCP That's true because I decided to check for separators ahead of time. Previous code was doing after the decode had returned multiple numbers.
> 
> So that's a toss up and this behavior was only added in 1.5.0 so which behavior is more correct?
> 
> See #57

I've checked the 1.4.1 version and following sample produces array of 0 length:

```c#
var hashids = new HashidsNet.Hashids("salt", minHashLength: 0, alphabet: "0123456789ABCDEF");

var hash = hashids.EncodeLong(1000);
Console.WriteLine(hash);

var decodedArray = hashids.DecodeLong(hash + "1");
Console.WriteLine(decodedArray.Length); // <-- 0 (No results decoded)
```

While on the PR branch this gives a MultipleResultsException.
Generally speaking, I support such perf changes as long as they mirror old behavior, otherwise, it would introduce a breaking change.