@manigandham 2022-04-05
@jetersen 

Can you please rebase on master to use the latest tests?

Also agree with @KeterSCP that the behavior shouldn't change. If there are no results decoded then it should not be returning a MultipleEntries exception. Is your code finding multiple integers in that hash?