@jetersen 2022-04-05
I took your example @KeterSCP 

You specifically added a one and with your setup tat is one of the separators:

![image](https://user-images.githubusercontent.com/1661688/161711212-ddcd3437-8d52-413b-b923-8e162043e3fa.png)

So how can `DecodeSingle` be wrong?