@KeterSCP 2022-04-05
> I took your example @KeterSCP 
> 
> You specifically added a `1` to decode and with your setup that is one of the separators:
> 
> ![image](https://user-images.githubusercontent.com/1661688/161711212-ddcd3437-8d52-413b-b923-8e162043e3fa.png)
> 
> So how can `DecodeSingle` be wrong?
> 
> I can remove the check for separators but than I should also remove the `MultipleEntriesException` cause there is no way to detect other than to look for separators after the result does not match the `GenerateHashFrom` equals check.
> 
> I tried to keep the behavior of `DecodeSingleLong` as it was coded in 1.5.0:
> https://github.com/ullmark/hashids.net/blob/b2c78773d05fcf8c9edc643f87db31e98554f2f0/src/Hashids.net/Hashids.cs#L180-L181

Well, i've just checked how it behaves for invalid hashes and compared how it did before. If such change is ok for the author of library, then i suppose it can be merged. Code looks good.
cc: @ullmark