@jetersen 2022-04-05
I created 535df3d as another way. This would result in NoResultException when trying to parse multiple when using `DecodeSingle`